Converts one tsx file to jsx
