+++ app/components/navbar/CollapsableNavbar.tsx
@@ -1,8 +1,8 @@
 import {memo, useRef} from "react";
 import {motion, useCycle} from "framer-motion";
 import {useDimensions} from "@/lib/useDimensions";
-import {Toggle} from "@/app/components/navbar/Toggle";
 import CollapsableNavbarItem from "@/app/components/navbar/CollapsableNavbarItem";
+import {Toggle} from "@/app/components/navbar/Toggle";
 
 const sidebar = {
     open: (height = 1000) => ({
